requirements.txt
numpy==1.18.1
Pillow==7.0.0
Flask==1.1.1
